requirements.txt
farm-haystack
chainlit
groq
tensorflow